backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.1
fastcore==1.8.5
ghapi==1.0.6
gitingest==0.1.4
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.47.1
tiktoken==0.9.0
tomli==2.2.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
httpx==0.28.1

# Add PyJWT for GitHub App authentication
PyJWT==2.8.0
cryptography==41.0.7  # For JWT signing with RSA keys


httpx
